updates to nevpt tests
