--- tests/nevpt2/22-nevpt2-h2o-triplet-rohf.ref
@@ -1,116 +1,0 @@
-converged SCF energy = -75.8159702269242
-SCF energy: -75.815970
-
-CASCI E = -75.8159702269242  E(CI) = -1.32331938692104  S^2 = 2.0000000
-CASCI E = -75.8159702269242  E(CI) = -1.32331938692104  S^2 = 2.0000000
-NEVPT2 test
-
-Computing NEVPT energies...
-
-Method:                                            nevpt2
-Method type:                                       ss
-Nuclear repulsion energy:                          9.168193300756
-Number of electrons:                               10
-Number of basis functions:                         41
-Reference wavefunction type:                       casci
-Number of reference states:                        1
-Number of reference microstates:                   1
-Number of frozen orbitals:                         0
-Number of core orbitals:                           4
-Number of active orbitals:                         2
-Number of external orbitals:                       35
-Number of active electrons:                        [(2, 0)]
-Spin–orbit coupling:                               None
-G-tensor:                                          False
-Reference density fitting?                         True
-Correlation density fitting?                       True
-Temporary directory path:                          /scratch/local/odhiambo.5
-
-Internal contraction:                              Full (= Partial)
-Compute singles amplitudes?                        False
-Overlap truncation parameter (singles):            1.00e-06
-Overlap truncation parameter (doubles):            1.00e-10
-
-Einsum Backend:                                    opt_einsum
-
-Transforming integrals to MO basis...
-Correlation energy [0]:                           -0.103528399236
-
-Computing energy of state #1...
-Reference state active-space energy:              -1.323319386921
-Reference state spin multiplicity:                 3
-Number of active electrons:                        [(2, 0)]
-Correlation energy [0']:                          -0.012118950320
-Correlation energy [+1']:                         -0.000000000000
-Correlation energy [-1']:                         -0.000000000000
-Correlation energy [+1]:                          -0.012444618488
-Correlation energy [-1]:                          -0.048131054020
-Correlation energy [+2]:                           0.000000000000
-Correlation energy [-2]:                          -0.001534337525
-Correlation energy [0]:                           -0.103528399236
-CASCI reference state total energy:              -75.815970226924
-NEVPT2 correlation energy:                        -0.177757359589
-Total NEVPT2 energy:                             -75.993727586513
-
-Summary of results for the SS-NEVPT2 calculation with the CASCI reference:
-------------------------------------------------------------------------------------------------------------------
-  State    Degen.        E(total)            dE(a.u.)        dE(eV)      dE(nm)       dE(cm-1)      Osc Str.  
-------------------------------------------------------------------------------------------------------------------
-    1        3          -75.993727586513     0.00000000       0.0000                      0.0000               
-----------------------------------------------------------------------------------------------------------------
-> CPU time for total NEVPT2 calculation      4.24 sec, wall time      0.18 sec
-
-Computing NEVPT energies...
-
-Method:                                            nevpt2
-Method type:                                       ss
-Nuclear repulsion energy:                          9.168193300756
-Number of electrons:                               10
-Number of basis functions:                         41
-Reference wavefunction type:                       casci
-Number of reference states:                        1
-Number of reference microstates:                   1
-Number of frozen orbitals:                         0
-Number of core orbitals:                           4
-Number of active orbitals:                         2
-Number of external orbitals:                       35
-Number of active electrons:                        [(2, 0)]
-Spin–orbit coupling:                               None
-G-tensor:                                          False
-Reference density fitting?                         True
-Correlation density fitting?                       False
-Temporary directory path:                          /scratch/local/odhiambo.5
-
-Internal contraction:                              Full (= Partial)
-Compute singles amplitudes?                        False
-Overlap truncation parameter (singles):            1.00e-06
-Overlap truncation parameter (doubles):            1.00e-10
-
-Einsum Backend:                                    opt_einsum
-
-Transforming integrals to MO basis...
-Correlation energy [0]:                           -0.103560001974
-
-Computing energy of state #1...
-Reference state active-space energy:              -1.323319386921
-Reference state spin multiplicity:                 3
-Number of active electrons:                        [(2, 0)]
-Correlation energy [0']:                          -0.012117824429
-Correlation energy [+1']:                         -0.000000000001
-Correlation energy [-1']:                         -0.000000003416
-Correlation energy [+1]:                          -0.012446056684
-Correlation energy [-1]:                          -0.048106110835
-Correlation energy [+2]:                           0.000000000000
-Correlation energy [-2]:                          -0.001534068340
-Correlation energy [0]:                           -0.103560001974
-CASCI reference state total energy:              -75.815970226924
-NEVPT2 correlation energy:                        -0.177764065678
-Total NEVPT2 energy:                             -75.993734292602
-
-Summary of results for the SS-NEVPT2 calculation with the CASCI reference:
-------------------------------------------------------------------------------------------------------------------
-  State    Degen.        E(total)            dE(a.u.)        dE(eV)      dE(nm)       dE(cm-1)      Osc Str.  
-------------------------------------------------------------------------------------------------------------------
-    1        3          -75.993734292602     0.00000000       0.0000                      0.0000               
-----------------------------------------------------------------------------------------------------------------
-> CPU time for total NEVPT2 calculation     21.09 sec, wall time      0.54 sec
